+++ src/App.jsx
@@ -5,10 +5,10 @@ import Register from './pages/register';
 import Login from './pages/login';
 import AddConsole from './pages/AddConsole';
 
-import ProtectedRoutes from './components/protetedRoutes';
+import ProtectedRoutes from './components/users/protetedRoutes';
 import Layout from './components/Layout';
 import Reservations from './pages/Reservations';
-import ConsoleDetail from './components/details/ConsoleDetail';
+import ConsoleDetail from './components/consoles/ConsoleDetail';
 
 import store from './redux/store';
 import './App.css';
