+++ src/components/consoles/ConsoleDetail.jsx
