+++ src/components/users/protetedRoutes.jsx
@@ -1,11 +1,11 @@
 import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
 import { Navigate } from 'react-router-dom';
-import { getToken } from '../utils/localStorage';
+import { getToken } from '../../utils/localStorage';
 
 const ProtectedRoutes = ({ children }) => {
   const { isAuthenticated } = useSelector((state) => state.user);
-  const token = getToken()
+  const token = getToken();
   const isAuthenticatedPersisted = isAuthenticated || !!token;
   if (!isAuthenticatedPersisted) {
     return <Navigate to="/login" />;
