+++ src/pages/login.jsx
@@ -1,4 +1,4 @@
-import LoginForm from '../components/login';
+import LoginForm from '../components/users/login';
 
 const Login = () => <LoginForm />;
